Moved non-exported methods to a separate file
srp.go was getting just a bit too large for me to navigate nicely.
Perhaps if I knew how to make better use of my editor (what is the
vscode equivalent of ctags?) I wouldn't have needed to do this.
